test(ui): move trace grid contracts to bun
