+++ docs/internal/development/plans/ui-test-latency/ui-test-coverage-implementation-plan.md
@@ -701,8 +701,12 @@ request behavior, or full dashboard composition until they are decomposed.
 
 ### Current component latency inventory
 
-The post-PR #1316 local profile groups the remaining dominant component cost by
-cause rather than runner alone:
+The post-PR #1331 local profile groups the remaining dominant component cost by
+cause rather than runner alone. The largest feature clusters in that profile
+were factory-session detail (3,586ms across 19 files), current selection
+(3,567ms across 32 files), workflow activity (3,511ms across 18 files),
+dashboard (1,163ms across 8 files), trace drill-down (1,032ms across 5 files),
+and header (899ms across 4 files).
 
 | Category | Representative file | Profiled source time | Reconciliation |
 | --- | --- | ---: | --- |
@@ -712,6 +716,7 @@ cause rather than runner alone:
 | Omnibus widget rendering | `current-selection-widget.provider-session-selection.test.tsx` | 720ms | Removed the redundant 500-line widget suite; inference-history selection state now runs in the feature-owned `useSelectedProviderSessionState` Bun contract (about 164ms), while selection controls remain covered by WorkItemCard and workstation-attempt contracts. |
 | Graph/editor host composition | `react-flow-current-activity-card-host-contract.test.tsx` | 700ms | Retain the small host boundary; continue replacing aggregate imports with owner modules. |
 | Workflow-activity bento aggregation | `workflow-activity-bento-card.test.tsx` | 385ms | Replaced the six-case, module-mocked graph omnibus with a two-case Bun contract around the extracted bento shell. Shared card chrome, editor controls, graph layout, and accessibility behavior stay with their existing feature owners. |
+| Trace-grid aggregation | `trace-grid-card.test.tsx` plus three scroll suites | 401ms for the Vitest omnibus, plus duplicate Bun work | Replaced the mocked graph-wide suite with a three-case Bun owner contract for states, table/selection behavior, and localization. Consolidated seven overlapping scroll cases into one long-table ownership invariant under `components/trace-grid-card/`. |
 | Cross-owner dashboard wiring | `dashboard-replay-wiring.component.test.tsx` | 697ms | Removed the app-shell duplicate: stream append and fixed-tick behavior are owned by the Bun snapshot hook, timeline store, slider, and bento snapshot contracts. |
 | Graph validation and charts | workflow editor validation and work chart tests | 675–691ms | WorkChart's 18 rendered contracts run under Bun. The duplicate workflow-activity editor validation suite is removed, and the seven feature-owned `useFactoryValidation` contracts now run under Bun in about 654ms with an injected validator rather than a process-global module mock. |
 | Chart wrapper duplication | `d3-information-card.test.tsx` | 540ms | Replaced the 600-line, ten-case Vitest suite with a three-case Bun owner contract for frame composition, localization/series wiring, and state forwarding. Legend, geometry, zoom, and state behavior remain with the dedicated WorkChart contracts. |
@@ -727,11 +732,10 @@ widget mounts, and process-global module mocks. Assertion execution is not the
 primary cost. Migration therefore starts by narrowing ownership and dependency
 seams; changing only the runner for an app-sized test preserves most latency.
 
-After the workflow-activity bento reconciliation, the complete component command
-owns 234 Bun files and 114 remaining Vitest files and completes locally in
-75.51s: 37.40s in Bun and 38.10s in Vitest. The last measured unit lane is
-18.05s, so the combined non-browser correctness feedback is approximately
-93.56s.
+After the trace-grid reconciliation, the complete component command owns 233
+Bun files and 113 remaining Vitest files and completes locally in 75.18s:
+37.64s in Bun and 37.54s in Vitest. The last measured unit lane is 18.05s, so
+the combined non-browser correctness feedback is approximately 93.23s.
 
 The same reconciliation removed the canonical-section omnibus widget loop.
 Every detail owner structurally composes `SelectionDetailLayout`; the base
